+++ src/components/ImageMapEditor/Console/index.ts
