+++ src/components/ImageMapEditor/ImageUploader/index.ts
