+++ src/components/ImageMapEditor/Main/Main.style.ts
