+++ src/components/ImageMapEditor/Main/index.ts
