+++ src/components/ImageMapEditor/RecordableImage/RecordItem.tsx
