+++ src/components/ImageMapEditor/RecordableImage/RecordableImage.tsx
