+++ src/components/ImageMapEditor/RecordableImage/index.ts
