+++ src/components/ImageMapEditor/useAppReducer.ts
